% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bws_access_token.R
\name{get_bws_access_token}
\alias{get_bws_access_token}
\title{Retrieve BWS Access Token from Environment}
\usage{
get_bws_access_token()
}
\value{
A character string containing the BWS access token.
}
\description{
Retrieves the \code{BWS_ACCESS_TOKEN} from environment variables. This token is
used to authenticate with the Bitwarden Secrets Manager API.
}
