% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bws_server_url.R
\name{get_bws_server_url}
\alias{get_bws_server_url}
\title{Retrieve BWS Server URL from Environment}
\usage{
get_bws_server_url()
}
\value{
A character string containing the BWS server URL.
}
\description{
Retrieves the \code{BWS_SERVER_URL} from environment variables. This URL is
used to define the Bitwarden Secrets Manager server endpoint.
}
