% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bws_secret.R
\name{get_bws_access_token}
\alias{get_bws_access_token}
\title{Get the bws_access_token from the environment variable}
\usage{
get_bws_access_token()
}
\value{
the bws_access_token as string
}
\description{
Get the bws_access_token from the environment variable
}
