% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bws_secret.R
\name{get_bws_secret}
\alias{get_bws_secret}
\title{Get a secret from you bws vault}
\usage{
get_bws_secret(
  key,
  access_token = Sys.getenv("BITWARDEN_ACCESs_TOKEN"),
  server_url = "https://vault.bitwarden.eu"
)
}
\arguments{
\item{key}{A character string}

\item{access_token}{Optional. A character string}

\item{server_url}{Optional. A character string}
}
\value{
A character string containing your secret
}
\description{
This function takes a key of your secret and returns the corresponding value
}
