% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bws_secret.R
\name{get_bws_server_url}
\alias{get_bws_server_url}
\title{Get the bws_server_url from the environment variable}
\usage{
get_bws_server_url()
}
\value{
the bws_server_url as string
}
\description{
Get the bws_server_url from the environment variable
}
